# Optimize WB2 by inlining matmuls and simplifying copies