With LLVM.

Before:

time bin/test_h5 smwb4 50 50 1e-3 100000
# of reps. = 100000
Residual = smwb4 50 9.60426e-07 2.31848e-11
ok -- cycle 50

real	0m3,743s
user	0m3,743s
sys	0m0,000s


After:

time bin/test_h5 smwb4 50 50 1e-3 100000
# of reps. = 100000
Residual = smwb4 50 9.60423e-07 2.31847e-11
ok -- cycle 50

real	0m2,266s
user	0m2,266s
sys	0m0,001s